requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=3.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
chromadb>=0.4.0
numpy>=1.21.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
celery>=5.2.0
redis>=4.0.0
flower>=1.0.0
jinja2>=3.0.0
python-multipart>=0.0.5
sentry-sdk